separate bundle dict test
