+++ AFQ/tests/test_api.py
@@ -26,7 +26,6 @@ from dipy.io.stateful_tractogram import StatefulTractogram, Space
 from dipy.testing.decorators import xvfb_it
 
 import AFQ.api.group as api
-import AFQ.api.bundle_dict as abd
 import AFQ.data as afd
 import AFQ.segmentation as seg
 import AFQ.utils.streamlines as aus
@@ -175,68 +174,6 @@ def create_dummy_bids_path(n_subjects, n_sessions, share_sessions=True):
             create_dummy_data(dmriprep_dir, subject)
 
     return bids_dir
-
-
-def test_BundleDict():
-    """
-    Tests bundle dict
-    """
-
-    # test defaults
-    afq_bundles = abd.BundleDict()
-
-    # bundles restricted within hemisphere
-    # NOTE: FA and FP cross midline so are removed
-    # NOTE: all others generate two bundles
-    num_hemi_bundles = (len(abd.BUNDLES)-2)*2
-
-    # bundles that cross the midline
-    num_whole_bundles = 2
-
-    assert len(afq_bundles) == num_hemi_bundles + num_whole_bundles
-
-    # Arcuate Fasciculus
-    afq_bundles = abd.BundleDict(["ARC"])
-
-    assert len(afq_bundles) == 2
-
-    # Forceps Minor
-    afq_bundles = abd.BundleDict(["FA"])
-
-    assert len(afq_bundles) == 1
-
-    # Cingulum Hippocampus
-    # not included but exists in templates
-    afq_bundles = abd.BundleDict(["HCC"])
-
-    assert len(afq_bundles) == 2
-
-    # Test "custom" bundle
-    afq_templates = afd.read_templates()
-    afq_bundles = abd.BundleDict({
-        "custom_bundle": {
-            "ROIs": [afq_templates["FA_L"],
-                     afq_templates["FP_R"]],
-            "rules": [True, True],
-            "cross_midline": False,
-            "uid": 1}})
-    afq_bundles.get("custom_bundle")
-
-    assert len(afq_bundles) == 1
-
-    # Vertical Occipital Fasciculus
-    # not included and does not exist in afq templates
-    with pytest.raises(
-            ValueError,
-            match="VOF_R is not in AFQ templates"):
-        afq_bundles = abd.BundleDict(["VOF"])
-        afq_bundles["VOF_R"]
-
-    afq_bundles = abd.BundleDict(["VOF"], seg_algo="reco80")
-    assert len(afq_bundles) == 2
-
-    afq_bundles = abd.BundleDict(["whole_brain"], seg_algo="reco80")
-    assert len(afq_bundles) == 1
 
 
 def test_AFQ_missing_files():
@@ -404,15 +341,6 @@ def test_AFQ_init():
                             npt.assert_equal(
                                 len(my_afq.wf_dict[sub][sess]),
                                 26)
-
-
-def test_AFQ_custom_bundle_dict():
-    bids_path = create_dummy_bids_path(3, 1)
-    bundle_dict = abd.BundleDict()
-    api.GroupAFQ(
-        bids_path,
-        preproc_pipeline="synthetic",
-        bundle_info=bundle_dict)
 
 
 @pytest.mark.nightly_basic
